Fix moving logic and fix tests for it
